+++ sevenn/nn/self_connection.py
@@ -1,3 +1,4 @@
+import torch
 import torch.nn as nn
 from e3nn.o3 import FullyConnectedTensorProduct, Irreps, Linear
 from e3nn.util.jit import compile_mode
@@ -58,9 +59,11 @@ class SelfConnectionIntro(nn.Module):
 
     def forward(self, data: AtomGraphDataType) -> AtomGraphDataType:
         assert self.fc_tensor_product is not None, 'Layer is not instantiated'
-        data[KEY.SELF_CONNECTION_TEMP] = self.fc_tensor_product(
-            data[self.key_x], data[self.key_operand]
-        )
+
+        x = data[self.key_x]
+        operand = data[self.key_operand]
+        data[KEY.SELF_CONNECTION_TEMP] = self.fc_tensor_product(x, operand)
+
         return data
 
 
@@ -104,7 +107,10 @@ class SelfConnectionLinearIntro(nn.Module):
 
     def forward(self, data: AtomGraphDataType) -> AtomGraphDataType:
         assert self.linear is not None, 'Layer is not instantiated'
-        data[KEY.SELF_CONNECTION_TEMP] = self.linear(data[self.key_x])
+
+        x = data[self.key_x]
+        data[KEY.SELF_CONNECTION_TEMP] = self.linear(x)
+
         return data
 
 
@@ -123,6 +129,10 @@ class SelfConnectionOutro(nn.Module):
         self.key_x = data_key_x
 
     def forward(self, data: AtomGraphDataType) -> AtomGraphDataType:
-        data[self.key_x] = data[self.key_x] + data[KEY.SELF_CONNECTION_TEMP]
+
+        x = data[self.key_x]
+        sc_temp = data[KEY.SELF_CONNECTION_TEMP]
+
+        data[self.key_x] = x + sc_temp
         del data[KEY.SELF_CONNECTION_TEMP]
         return data