+++ CHANGELOG.md
@@ -4,10 +4,18 @@ All notable changes to this project will be documented in this file.
 
 ## [0.12.0-dev]
 ### Added
+- LAMMPS-MLIAP integration with GhostExchangeOp
+
+### Added
 - Example config for fine-tuning the SevenNet-MF-ompa model
+- FlashTP support (https://github.com/SNU-ARC/flashTP)
+
+### Changed
+- ninja dependency added
 
 ### Fixed
 - Multi modal model finetuning: Pass modal_map into config during multi-fidelity continual training #232
+- omat deploy parallel fixed
 
 
 ## [0.11.2]